TBLA: split into main component file and utils
