+++ src/index.js
@@ -13,8 +13,12 @@ import {
   WebGLRenderer,
 } from 'three'
 import { MeshLine, MeshLineMaterial } from 'three.meshline'
+import {
+  hue2rgb,
+  setHSL,
+  getNearbyObjectsFor,
+} from './utils'
 
-const CONNECT_DISTANCE = 400
 const MAX_CONNECT_LINES = 4
 const NUM_SPHERES = 100
 const SPHERE_DIAMETER = 40
@@ -26,37 +30,9 @@ const camera = new PerspectiveCamera(55, window.innerWidth / window.innerHeight,
 const scene = new Scene()
 const spheres = new Group()
 const lineMaterial = new MeshLineMaterial({
-    lineWidth: LINE_WIDTH,
-    color: new Color('purple'),
-  })
-
-const hue2rgb = ( p, q, t ) => {
-  if ( t < 0 ) t += 1;
-  if ( t > 1 ) t -= 1;
-  if ( t < 1 / 6 ) return p + ( q - p ) * 6 * t;
-  if ( t < 1 / 2 ) return q;
-  if ( t < 2 / 3 ) return p + ( q - p ) * 6 * ( 2 / 3 - t );
-  return p;
-}
-
-const setHSL = ( h, s, l ) => {
-  let r,g,b
-  // h,s,l ranges are in 0.0 - 1.0
-  h = ( ( h % 1 ) + 1 ) % 1 // _Math.euclideanModulo( h, 1 );
-  s = Math.max( 0, Math.min( 1, s ) )//_Math.clamp( s, 0, 1 );
-  l = Math.max( 0, Math.min( 1, l ) )//_Math.clamp( l, 0, 1 );
-
-  if ( s === 0 ) {
-    r = g = b = l;
-  } else {
-    var p = l <= 0.5 ? l * ( 1 + s ) : l + s - ( l * s );
-    var q = ( 2 * l ) - p;
-    r = hue2rgb( q, p, h + 1 / 3 );
-    g = hue2rgb( q, p, h );
-    b = hue2rgb( q, p, h - 1 / 3 );
-  }
-  return new Color(r,g,b);
-}
+  lineWidth: LINE_WIDTH,
+  color: new Color('purple'),
+})
 
 const addSpheres = () => {
   for (var i = 0; i < NUM_SPHERES; i++) {
@@ -69,19 +45,6 @@ const addSpheres = () => {
     spheres.add(sphere)
   }
   scene.add(spheres)
-}
-
-const getNearbyObjectsFor = (obj, group) => {
-  let objs = []
-  for (var j = 0; j < group.children.length; j++) {
-    if (obj.uuid == group.children[j].uuid) continue
-
-    var dist = obj.position.distanceTo(group.children[j].position)
-    if (dist < CONNECT_DISTANCE) {
-      objs.push(group.children[j])
-    }
-  }
-  return objs
 }
 
 const init = container => {
@@ -124,35 +87,34 @@ const init = container => {
 }
 
 const render = () => {
+  const time = Date.now() * 0.00005;
+  //optional: moving around mouse changes the camera animation
+  //camera.position.x += ( mouseX - camera.position.x ) * 0.05;
+  //camera.position.y += ( - mouseY - camera.position.y ) * 0.03;
+  //camera.position.x = camera.position.x + 1
+  //camera.lookAt( scene.position );
+  const h = ( 360 * ( 1.0 + time ) % 360 ) / 360;
+  sphereMaterial.color.setHSL( h, 0.5, 0.5 );
+  lineMaterial.uniforms.color.value = setHSL(h, 0.5, 0.5)
 
-        var time = Date.now() * 0.00005;
-        //optional: moving around mouse changes the camera animation
-        //camera.position.x += ( mouseX - camera.position.x ) * 0.05;
-        //camera.position.y += ( - mouseY - camera.position.y ) * 0.03;
-        //camera.lookAt( scene.position );
-        const h = ( 360 * ( 1.0 + time ) % 360 ) / 360;
-        sphereMaterial.color.setHSL( h, 0.5, 0.5 );
-        lineMaterial.uniforms.color.value = setHSL(h, 0.5, 0.5)
+  renderer.render( scene, camera );
+}
 
-        renderer.render( scene, camera );
-      }
+const animate = () => {
+  requestAnimationFrame(animate)
+  render()
+}
 
 export default class extends Component {
 
   componentDidMount() {
     init(this.container)
-      function animate() {
-        requestAnimationFrame( animate );
-        render();
-      }
-        requestAnimationFrame( animate );
-        render();
+    animate()
   }
 
   render() {
     return (
-      <div ref={c => this.container = c}>
-      </div>
+      <div ref={c => this.container = c} />
     )
   }
 }
